Put func decls in corresponding header file
Also add 'static' to applicable functions.
